backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pymongo
python-jose
passlib[bcrypt]
langchain
langchain-openai
pytest
email-validator
